TECH-287: Fixed i18n, updated shared libs
