+++ package.json
@@ -36,12 +36,12 @@
     },
     "dependencies": {
         "@apollo/react-hooks": "^3.1.5",
-        "@jahia/data-helper": "^0.0.26",
+        "@jahia/data-helper": "^1.0.0",
         "@jahia/design-system-kit": "^1.1.8",
         "@jahia/icons": "^1.1.1",
         "@jahia/moonstone": "^0.17.4",
-        "@jahia/react-material": "^2.1.1",
-        "@jahia/ui-extender": "^0.5.4",
+        "@jahia/react-material": "^3.0.1",
+        "@jahia/ui-extender": "^1.0.0",
         "@material-ui/core": "^3.9.3",
         "@material-ui/icons": "^3.0.1",
         "@material-ui/lab": "^3.0.0-alpha.30",
