+++ src/javascript/Edit/EditContent.action.jsx
@@ -4,9 +4,11 @@ import {useNodeChecks} from '@jahia/data-helper';
 import * as PropTypes from 'prop-types';
 import {useSelector} from 'react-redux';
 import {Constants} from '~/ContentEditor.constants';
+import {useTranslation} from "react-i18next";
 
 export const EditContent = ({path, render: Render, loading: Loading, ...otherProps}) => {
     const {redirect} = useContentEditorHistory();
+    useTranslation('content-editor');
     const {language} = useSelector(state => ({language: state.language}));
     const res = useNodeChecks(
         {path: path, language: language},
